smarthome/breakout_garden/requirements.txt
bme680
prometheus-client
dotenv
pimoroni-sgp30
veml6075
smbus2